+++ packages/types/src/interfaces/runtime/definitions.ts
@@ -52,6 +52,7 @@ export default {
       block: 'Block',
       justification: 'Justification'
     },
+    SignerPayload: 'GenericExtrinsicSignerPayload',
     ValidatorId: 'AccountId',
     Weight: 'u32',
     WeightMultiplier: 'Fixed64',
