+++ packages/types/src/types.ts
@@ -216,3 +216,84 @@ export interface IExtrinsic extends IExtrinsicSignable<IExtrinsic>, ExtrinsicSig
   readonly type: number;
   readonly version: number;
 }
+
+export interface SignerPayloadJSON {
+  /**
+   * @description The ss-58 encoded address
+   */
+  address: string;
+
+  /**
+   * @description The checkpoint hash of the block, in hex
+   */
+  blockHash: string;
+
+  /**
+   * @description The checkpoint block number, in hex
+   */
+  blockNumber: string;
+
+  /**
+   * @description The era for this transaction, in hex
+   */
+  era: string;
+
+  /**
+   * @description The genesis hash of the chain, in hex
+   */
+  genesisHash: string;
+
+  /**
+   * @description The encoded method (with arguments) in hex
+   */
+  method: string;
+
+  /**
+   * @description The nonce for this transaction, in hex
+   */
+  nonce: string;
+
+  /**
+   * @description The current spec version for  the runtime
+   */
+  specVersion: string;
+
+  /**
+   * @description The tip for this transaction, in hex
+   */
+  tip: string;
+
+  /**
+   * @description The version of the extrinsic we are dealing with
+   */
+  version: number;
+}
+
+export interface SignerPayloadRawBase {
+  /**
+   * @description The hex-encoded data for this request
+   */
+  data: string;
+
+  /**
+   * @description The type of the contained data
+   */
+  type?: 'bytes' | 'payload';
+}
+
+export interface SignerPayloadRaw extends SignerPayloadRawBase {
+  /**
+   * @description The ss-58 encoded address
+   */
+  address: string;
+
+  /**
+   * @description The type of the contained data
+   */
+  type: 'bytes' | 'payload';
+}
+
+export interface ISignerPayload {
+  toPayload (): SignerPayloadJSON;
+  toRaw (): SignerPayloadRaw;
+}
